# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "JSON error response structure",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "IssueFilter": {
        "description": "Filter for issues listing",
        "enum": [
          "open",
          "resolved",
          "muted",
          "all"
        ],
        "type": "string"
      },
      "IssueResponse": {
        "description": "Response for API",
        "properties": {
          "event_count": {
            "format": "int32",
            "type": "integer"
          },
          "first_seen": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "is_muted": {
            "type": "boolean"
          },
          "is_resolved": {
            "type": "boolean"
          },
          "last_seen": {
            "format": "date-time",
            "type": "string"
          },
          "level": {
            "nullable": true,
            "type": "string"
          },
          "platform": {
            "nullable": true,
            "type": "string"
          },
          "project_id": {
            "format": "int32",
            "type": "integer"
          },
          "short_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "short_id",
          "title",
          "value",
          "first_seen",
          "last_seen",
          "event_count",
          "is_resolved",
          "is_muted"
        ],
        "type": "object"
      },
      "IssueSort": {
        "description": "Sort mode for issues listing",
        "enum": [
          "digest_order",
          "last_seen"
        ],
        "type": "string"
      },
...
      "OffsetPaginatedIssueResponse": {
        "description": "Paginated issue response (offset-based) - concrete type for OpenAPI",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int64",
            "type": "integer"
          },
          "per_page": {
            "format": "int64",
            "type": "integer"
          },
          "total_count": {
            "format": "int64",
            "type": "integer"
          },
          "total_pages": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total_count",
          "page",
          "per_page",
          "total_pages"
        ],
        "type": "object"
      },
...
      "SortOrder": {
        "description": "Sort order direction",
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
...
      "UpdateIssueState": {
        "description": "Request to update issue state",
        "properties": {
          "is_muted": {
            "nullable": true,
            "type": "boolean"
          },
          "is_resolved": {
            "nullable": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "hex",
        "description": "40-character hex API token",
        "scheme": "bearer",
        "type": "http"
      },
      "sentry_auth": {
        "description": "Sentry SDK authentication header or ?sentry_key= query parameter",
        "in": "header",
        "name": "X-Sentry-Auth",
        "type": "apiKey"
      },
      "session_auth": {
        "in": "cookie",
        "name": "rustrak_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Rustrak Team"
    },
    "description": "Ultra-lightweight error tracking server compatible with Sentry SDKs",
    "license": {
      "name": "GPL-3.0"
    },
    "title": "Rustrak API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/projects/{project_id}/issues": {
      "get": {
        "operationId": "list_issues",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Page number (1-indexed, default: 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Items per page (default: 20)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Sort mode (default: last_seen)",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IssueSort"
            }
          },
          {
            "description": "Sort order direction (default: desc)",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "description": "Filter: open (not resolved, not muted), resolved, muted, all",
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IssueFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffsetPaginatedIssueResponse"
                }
              }
            },
            "description": "List of issues"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/issues\nLists issues for a project with offset-based pagination",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/projects/{project_id}/issues/{issue_id}": {
      "delete": {
        "operationId": "delete_issue",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Issue ID",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Issue deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "DELETE /api/projects/{project_id}/issues/{issue_id}\nSoft-deletes an issue",
        "tags": [
          "issues"
        ]
      },
      "get": {
        "operationId": "get_issue",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Issue ID",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            },
            "description": "Issue details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/issues/{issue_id}\nGets a single issue by ID",
        "tags": [
          "issues"
        ]
      },
      "patch": {
        "operationId": "update_issue",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Issue ID",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            },
            "description": "Issue updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "PATCH /api/projects/{project_id}/issues/{issue_id}\nUpdates issue state (resolve, mute, etc.)",
        "tags": [
          "issues"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer_auth": []
    },
    {
      "session_auth": []
    },
    {
      "sentry_auth": []
    }
  ],
...
}